refactor: create render card function
