+++ src/views/profile/bookmarks-view/index.tsx
@@ -1,16 +1,10 @@
 import { useAllBookmarks } from 'hooks/bookmarks'
 import BaseNewLayout from 'layouts/base-new'
 import AuthenticatedView from 'views/authenticated-view'
-import { getTutorialSlug } from 'views/collection-view/helpers'
-import { formatTutorialData } from 'lib/learn-client/api/tutorial/formatting'
-import { ProductUsed } from 'lib/learn-client/types'
-import { ApiBookmark } from 'lib/learn-client/api/api-types'
-import TutorialCard from 'components/tutorial-card'
-import getReadableTime from 'components/tutorial-meta/components/badges/helpers'
 import CardsGridList from 'components/cards-grid-list'
 import Text from 'components/text'
 import Heading from 'components/heading'
-import { BookmarkButtonIconOnly } from 'components/bookmark-button'
+import renderBookmarkCard from './helpers/render-bookmark-cards'
 import s from './bookmarks-view.module.css'
 
 /**
@@ -53,46 +47,7 @@ const ProfileBookmarksViewContent = () => {
 				Your Bookmarks
 			</Heading>
 			<CardsGridList isOrdered={false}>
-				{bookmarks?.map((bookmark: ApiBookmark) => {
-					const {
-						id,
-						slug,
-						description,
-						readTime,
-						handsOnLab,
-						video,
-						name,
-						collectionCtx,
-						productsUsed,
-					} = formatTutorialData(bookmark.tutorial)
-					const duration = getReadableTime(readTime)
-					return (
-						<li key={id}>
-							<TutorialCard
-								heading={name}
-								description={description}
-								url={getTutorialSlug(slug, collectionCtx.default.slug)}
-								duration={duration}
-								hasInteractiveLab={Boolean(handsOnLab)}
-								hasVideo={Boolean(video)}
-								productsUsed={productsUsed.map(
-									(p: ProductUsed) => p.product.slug
-								)}
-								eyebrowSlot={
-									<>
-										<span>{duration}</span>
-										<BookmarkButtonIconOnly
-											isBookmarked={true}
-											handleOnClick={() =>
-												console.log('Trigger confirmation dialog!')
-											}
-										/>
-									</>
-								}
-							/>
-						</li>
-					)
-				})}
+				{bookmarks?.map(renderBookmarkCard)}
 			</CardsGridList>
 		</div>
 	)
